pyproject.toml
[project]
name = "gen-aac"
version = "0.1.0"
description = "Generating AAC Symbol using Generative AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-firestore>=2.21.0",
    "google-cloud-storage>=3.6.0",
    "google-genai>=1.53.0",
    "httpx>=0.28.1",
    "litellm>=1.80.5",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["genaac"]
